Fix circular ref; free connection before reconnect
Need to weaken the $self ref so AE watchers don't hold references at exit.

Need to undef the connection so reconnecting doesn't spawn a duplicate.

There's probably still a bug here. The stream was getting doubled, and
tripled, sometimes. Which likely means there were recoverable errors that
should not result in a reconnect. At least with this change, a reconnect
should result in a single stream watcher.